Pagina inicio
se crea una nueva pagina inicio para el proyecto
