+++ src/app/inicio/inicio-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { InicioPage } from './inicio.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: InicioPage
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class InicioPageRoutingModule {}
