+++ src/app/inicio/inicio.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+
+import { IonicModule } from '@ionic/angular';
+
+import { InicioPageRoutingModule } from './inicio-routing.module';
+
+import { InicioPage } from './inicio.page';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    FormsModule,
+    IonicModule,
+    InicioPageRoutingModule
+  ],
+  declarations: [InicioPage]
+})
+export class InicioPageModule {}
