+++ src/app/inicio/inicio.page.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-inicio',
+  templateUrl: './inicio.page.html',
+  styleUrls: ['./inicio.page.scss'],
+})
+export class InicioPage implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
